+++ src/Control/Effect.hs
@@ -2,7 +2,7 @@ module Control.Effect
 ( module X
 ) where
 
-import Control.Effect.Carrier   as X (Carrier, Effect)
+import Control.Effect.Carrier   as X ((:+:), Carrier, Effect, Member)
 import Control.Effect.Cull      as X (Cull, CullC, OnceC)
 import Control.Effect.Cut       as X (Cut, CutC)
 import Control.Effect.Error     as X (Error, ErrorC)
@@ -16,6 +16,5 @@ import Control.Effect.Reader    as X (Reader, ReaderC)
 import Control.Effect.Resource  as X (Resource, ResourceC)
 import Control.Effect.Resumable as X (Resumable, ResumableC, ResumableWithC)
 import Control.Effect.State     as X (State, StateC)
-import Control.Effect.Sum       as X ((:+:), Member)
 import Control.Effect.Trace     as X (Trace, TraceByPrintingC, TraceByIgnoringC, TraceByReturningC)
 import Control.Effect.Writer    as X (Writer, WriterC)