+++ src/Control/Effect/Cut.hs
@@ -14,7 +14,6 @@ module Control.Effect.Cut
 import Control.Applicative (Alternative(..))
 import Control.Effect.Carrier
 import Control.Effect.NonDet
-import Control.Effect.Sum
 import Control.Monad (MonadPlus(..))
 import Control.Monad.Fail
 import Control.Monad.Fix