+++ src/Control/Effect/Error.hs
@@ -11,7 +11,6 @@ module Control.Effect.Error
 
 import Control.Applicative (Alternative(..), liftA2)
 import Control.Effect.Carrier
-import Control.Effect.Sum
 import Control.Monad (MonadPlus(..), (<=<))
 import Control.Monad.Fail
 import Control.Monad.Fix