+++ src/Control/Effect/Fresh.hs
@@ -12,7 +12,6 @@ module Control.Effect.Fresh
 import Control.Applicative (Alternative(..))
 import Control.Effect.Carrier
 import Control.Effect.State
-import Control.Effect.Sum
 import Control.Monad (MonadPlus(..))
 import Control.Monad.Fail
 import Control.Monad.Fix