+++ src/Control/Effect/Interpose.hs
@@ -13,7 +13,6 @@ module Control.Effect.Interpose
 import Control.Applicative
 import Control.Effect.Carrier
 import Control.Effect.Reader
-import Control.Effect.Sum
 import Control.Monad (MonadPlus (..))
 import Control.Monad.Fail
 import Control.Monad.Fix