+++ src/Control/Effect/Lift.hs
@@ -10,7 +10,6 @@ module Control.Effect.Lift
 
 import Control.Applicative (Alternative(..))
 import Control.Effect.Carrier
-import Control.Effect.Sum
 import Control.Monad (MonadPlus(..))
 import Control.Monad.Fail
 import Control.Monad.Fix