+++ src/Control/Effect/NonDet.hs
@@ -11,7 +11,6 @@ module Control.Effect.NonDet
 
 import Control.Applicative (Alternative(..))
 import Control.Effect.Carrier
-import Control.Effect.Sum
 import Control.Monad (MonadPlus(..))
 import Control.Monad.Fail
 import Control.Monad.Fix